Create The Frog's Way
In one very long and narrow pond, a frog jumps on water lilies. The water lilies in the pond are arranged in a single row. The frog starts jumping from the first water lily of the row and wants to finish on the last one. But due to the harmful nature of the frog, it agrees to jump only forward over one or two water lilies. For example, from water lily pad number 1, she can only jump to water lilies number 3 and water lily number 4.

Some water lilies have mosquitoes sitting on them. Namely, ai mosquitoes sit on the i-th water lily. When a frog lands on a water lily, it eats all the mosquitoes sitting on it. The frog wants to plan its route so that it eats as many mosquitoes as possible. Help her by telling her which water lilies she should visit on her way.

Input
The first entry line contains n — the number of water lilies in the pond (1 ≤ n ≤ 100,000). The second line contains n numbers separated by single spaces. The i-th number tells you how many mosquitoes are sitting on the i-th water lily (1 ≤ i ≤ n). All numbers are integers, non-negative, and do not exceed 1000.

Output
In the first line, print one number, which is the maximum number of mosquitoes that the frog can eat. In the second line, print the sequence of numbers, the numbers of the water lilies that the frog should visit, in ascending order. If there are several solutions, output any of them.
If the frog can't get to the last water lily, output one number −1.